nbs/test_apps/with_auth/requirements.txt
python-fasthtml
httpx
python-dotenv
PyJWT
